Menambahkan fitur pencarian di index.php

+++ SQL/Index.php
@@ -17,6 +17,12 @@
     <h1>Data Mahasiswa</h1>
     <a href="penambahan.php">Tambah Data Mahasiswa</a>
     <br><br>
+    <form>
+        <label for="pencarian">Cari Data</label>
+        <input type="text" name="cari" id="pencarian" size="30" value="<?php if (isset($_GET['cari'])) {echo $_GET['cari'];} ?>" autofocus>
+        <button type="submit">Cari</button>
+        <br><br>
+    </form>
     <table border="1" cellspacing="0" cellpadding="10">
         <tr>
             <th class="warna">No.</th>
@@ -26,24 +32,48 @@
             <th class="warna">Email</th>
             <th class="warna">Jurusan</th>
         </tr>
-        <?php $nomor = 1; /* Variabel untuk memberikan penomoran dari 1 */?>
-        <?php foreach ($mhs as $siswa) : ?>
-        <tr>
-            <td class="warna"><?= $nomor; ?></td>
-            <td>
-                <a href="update.php?id=<?= $siswa["Id"] ?>">Ubah</a>
-                <span>|</span>
-                <a href="hapus.php?id=<?= $siswa["Id"]; ?>" onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?') /*Konfirmasi sebelum menghapus datanya dengan js*/">Hapus</a>
-            </td>
-            <td>
-                <img src="../Img/<?= htmlspecialchars($siswa["Gambar"]); ?>" alt="Profil" width="50" height="50">
-            </td>
-            <td><?= htmlspecialchars($siswa["Nama"]); ?></td>
-            <td><?= htmlspecialchars($siswa["Email"]); ?></td>
-            <td><?= htmlspecialchars($siswa["Jurusan"]); ?></td>
-        </tr>
-        <?php $nomor++; ?>
-        <?php endforeach ?>
-    </table>
+        <?php if (!isset($_GET["cari"])) : ?>
+                <?php $nomor = 1; /* Variabel untuk memberikan penomoran dari 1 */?>
+                <?php foreach ($mhs as $siswa) : ?>
+                        <tr>
+                            <td class="warna"><?= $nomor; ?></td>
+                            <td>
+                                <a href="update.php?id=<?= $siswa["Id"] ?>">Ubah</a>
+                                <span>|</span>
+                                <a href="hapus.php?id=<?= $siswa["Id"]; ?>" onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?') /*Konfirmasi sebelum menghapus datanya dengan js*/">Hapus</a>
+                            </td>
+                            <td>
+                                <img src="../Img/<?= htmlspecialchars($siswa["Gambar"]); ?>" alt="Profil" width="50" height="50">
+                            </td>
+                            <td><?= htmlspecialchars($siswa["Nama"]); ?></td>
+                            <td><?= htmlspecialchars($siswa["Email"]); ?></td>
+                            <td><?= htmlspecialchars($siswa["Jurusan"]); ?></td>
+                        </tr>
+                <?php $nomor++; ?>
+                <?php endforeach; ?>
+            </table>
+        <?php else : ?>
+            <?php $no = 1; ?>
+            <?php foreach ($cari_data as $data) : ?>
+                <tr>
+                    <td class="warna"><?= $no; ?></td>
+                    <td>
+                        <a href="update.php?id=<?= $data["Id"] ?>">Ubah</a>
+                        <span>|</span>
+                        <a href="hapus.php?id=<?= $data["Id"]; ?>" onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?') /*Konfirmasi sebelum menghapus datanya dengan js*/">Hapus</a>
+                    </td>
+                    <td>
+                        <img src="../Img/<?= htmlspecialchars($data["Gambar"]); ?>" alt="Profil" width="50" height="50">
+                    </td>
+                    <td><?= htmlspecialchars($data["Nama"]); ?></td>
+                    <td><?= htmlspecialchars($data["Email"]); ?></td>
+                    <td><?= htmlspecialchars($data["Jurusan"]); ?></td>
+                </tr>
+            <?php $no++; ?>
+            <?php endforeach ?>
+                </table>
+            <br>
+            <a href="index.php">Kembali ke semua daftar mahasiswa</a>
+        <?php endif; ?>
 </body>
 </html>
